refactor data source with new functions
