+++ transip/data_source_domain.go
@@ -2,6 +2,7 @@ package transip
 
 import (
 	"fmt"
+
 	"github.com/hashicorp/terraform/helper/schema"
 	"github.com/transip/gotransip"
 	"github.com/transip/gotransip/domain"
@@ -36,15 +37,11 @@ func dataSourceDomainRead(d *schema.ResourceData, m interface{}) error {
 		return fmt.Errorf("Could not request domain info [%s]: %s", domainName, err)
 	}
 
-	d.Set("name", domainName)
-
-	nameServers := make([]string, 0)
-	if info.Nameservers != nil {
-		for _, ns := range info.Nameservers {
-			nameServers = append(nameServers, ns.Hostname)
-		}
+	if err := d.Set("name", domainName); err != nil {
+		return err
 	}
-	if err := d.Set("name_servers", nameServers); err != nil {
+
+	if err := d.Set("name_servers", flattenNameServers(info.Nameservers)); err != nil {
 		return err
 	}
