+++ transip/helpers/err_fmt/errors.go
@@ -1,4 +1,4 @@
-package transip
+package err_fmt
 
 import (
 	"fmt"
