+++ transip/helpers/err_fmt/errors_test.go
@@ -1,4 +1,4 @@
-package transip
+package err_fmt
 
 import (
 	"regexp"
